backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
numpy==2.1.1
scipy==1.13.1
python-dotenv==1.0.1

